Separa funções e adiciona pergunta de continuidade
